Break out asset test scaffolding into Resources
